refactor(onboarding): extract styles to OnboardingV2Styles component
Move ~1125 lines of <style jsx global> CSS (animations, keyframes,
feed limits, parallax, glassmorphism) to a separate component.
All class names are onb- prefixed, no collision risk.
